requirements.txt
# Face Recognition Attendance System
# Python 3.8+

dlib>=19.24
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Flask>=3.0.0
Pillow>=10.0.0
